Fix parsing header-only ADIF without EOF
This was already fixed in #6 for LotW files with the non-standard EOF headers, but in my app I was still pre-stripping the EOF and was crashing on a -1 array index at https://github.com/Matir/adifparser/blob/e70240124a37612ec1c6cc4968696cc2e6f58812/adifreader.go#L130. This should fix the general case of header-only ADIF files in addition to the special case of LotW EOF files.